Fix Master.Subscribe() event stream.
The Master Subscribe implementation has 2 problems.

First, the RecordIO reader does not correctly calculate the remaining
bytes to read. The reader was assuming bufio Reader.Read([]bytes)
returns the number of bytes remaining, but it actually returns the
number of bytes read.

Second, the same Event object is overwritten by each message that we
read to the event stream. If the caller is using a buffered channel to
read events, this can cause the event to be overwritten while the
caller is still consuming it.